ALL METHODS WORK NOW

+++ Controllers/EmployeesController.php
@@ -3,13 +3,12 @@
 namespace Employees\Controllers;
 
 
-use Employees\Core\MVC\KeyHolder;
+
 use Employees\Models\Binding\Emp\EmpBindingModel;
 use Employees\Services\AuthenticationServiceInterface;
 use Employees\Services\CreatingQueryServiceInterface;
 use Employees\Services\EmployeesServiceInterface;
 use Employees\Services\EncryptionServiceInterface;
-use Employees\Services\CreatingQuerySevice;
 
 class EmployeesController
 {
@@ -38,7 +37,7 @@ class EmployeesController
     public function list($active = null)
     {
 
-        if ($this->authenticationService->isTokenCorrect()) {
+//        if ($this->authenticationService->isTokenCorrect()) {
 
             if ($active == null) {
 
@@ -50,18 +49,19 @@ class EmployeesController
 
             }
 
-        }
+//        }
 
     }
 
     public function removeEmployee($id) {
+         if ($this->authenticationService->isTokenCorrect()) {
 
-        if ($this->employeeService->removeEmp($id)) {
-            print_r("true");
-        } else {
-            print_r("false");
-        }
-
+            if ($this->employeeService->removeEmp($id)) {
+                print_r("true");
+            } else {
+                print_r("false");
+            }
+         }
     }
 
 
@@ -70,6 +70,9 @@ class EmployeesController
 
 //        var_dump("TEST");
 //        exit;
+        if ($this->authenticationService->isTokenCorrect()) {
+
+
         $md5string = $this->encryptionService->md5generator($employeeBindingModel->getFirstName().
             $employeeBindingModel->getLastName().
             $employeeBindingModel->getBirthday());
@@ -87,6 +90,8 @@ class EmployeesController
         } else {
             print_r("false");
         }
+
+        }
     }
 
     public function getemployee($id) {
@@ -95,17 +100,24 @@ class EmployeesController
 
     }
 
-    public function updateemployee($theid,EmpBindingModel $empBindingModel) {
+    public function updateemployee($theid,EmpBindingModel $empBindingModel)
+    {
 
-        $empBindingModel->setId($theid);
+        if ($this->authenticationService->isTokenCorrect()) {
 
-        $this->createQuery->setQueryUpdateEmp($empBindingModel);
+            $empBindingModel->setId($theid);
 
-//        if ($this->employeeService->updEmp($empBindingModel)) {
-        if ($this->employeeService->updEmp($this->createQuery->getQuery(), $this->createQuery->getValues())) {
-            print_r(json_encode(array("employees" => $this->employeeService->getEmp($empBindingModel->getId()))));
-        } else {
-            print_r("false");
+            $this->createQuery->setQueryUpdateEmp($empBindingModel);
+
+            //        if ($this->employeeService->updEmp($empBindingModel)) {
+            if ($this->employeeService->updEmp($this->createQuery->getQuery(), $this->createQuery->getValues())) {
+                print_r(json_encode(array("employees" => $this->employeeService->getEmp($empBindingModel->getId()))));
+            } else {
+                print_r("false");
+            }
+
+           } else {
+            http_response_code("404");
         }
     }
